refactor: rename file utils and update import accordingly
